refactor: :art: componets folder
moved compinents to components folder
