+++ src/app/app-routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { ProductsContainerComponent } from './products-container/products-container.component';
-import { CartContainerComponent } from './cart-container/cart-container.component';
-import { ProductPageComponent } from './product-page/product-page.component';
+import { ProductsContainerComponent } from './components/products-container/products-container.component';
+import { CartContainerComponent } from './components/cart-container/cart-container.component';
+import { ProductPageComponent } from './components/product-page/product-page.component';
 
 const routes: Routes = [
   { path: '', component: ProductsContainerComponent },
