+++ src/app/components/product-page/product-page.component.scss
