+++ src/app/components/products-container/product-card/product-card.component.scss
